Small changes on Windows
* Fixed compiled-in file paths for signalflows unit test.
* Resolved compiler warnings and errors due to template type resolution ambiguities.

+++ src/apps/baseline_renderer/main.cpp
@@ -153,7 +153,7 @@ int main( int argc, char const * const * argv )
 
     // Assume a fixed length for the interpolation period.
     // Ideally, this roughly matches the update rate of the scene sender.
-    const std::size_t cInterpolationLength = std::max( 2048ul, periodSize );
+    const std::size_t cInterpolationLength = std::max( static_cast<std::size_t>(2048), periodSize );
 
     /* Set up the filter matrix for the diffusion filters. */
     std::size_t const diffusionFilterLength = 63; // fixed filter length of the filters in the compiled-in matrix
